+++ lib/src/network/core/ssl_pinning_helper.dart
@@ -1,0 +1,2 @@
+export 'ssl_pinning_helper_native.dart'
+    if (dart.library.html) 'ssl_pinning_helper_web.dart';